backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
sqlalchemy
psycopg2-binary
asyncpg
python-dotenv
python-multipart
pandas
httpx
requests

# AI & LangChain
groq
langchain
langchain-community
langchain-core
langchain-groq
langchain-classic
langchain-experimental

# Data
chromadb

# Utils
redis
tenacity
watchdog
python-dateutil

# Testing
pytest
pytest-asyncio
